move test cases into src
